# docs/openapi.yaml
openapi: 3.0.3
...
info:
  title: Prompt Patrol API
  version: 0.5.1
  description: |
    Instructor-facing decision-support API for flagging potentially AI-generated
    short answers in software engineering courses.

    ## Authentication model

    Prompt Patrol is an **OpenID Connect Relying Party**, not an OAuth client.
    It has no third-party resource server and requests no delegated authority.
    Entra ID answers exactly one question, once per login: *who is this person?*

    Consequence: **no access token, no refresh token, and no JWT is stored
    anywhere.** The ID token is validated, its claims extracted, and the token
    discarded. See IETF `draft-ietf-oauth-browser-based-apps` §7.1
    ("Single-Domain Browser-Based Applications").

    Session state is server-side. The browser holds an opaque 256-bit random
    identifier in a cookie; that value is a database lookup key, meaningless
    outside this system.

    The authority is **tenant-scoped**
    (`login.microsoftonline.com/{tenant-id}/...`), never `/common`. With
    `/common`, any Microsoft account could complete the flow, and a personal
    account with a colliding email address could claim a provisioned slot
    before the intended user's first login. See "Provisioning" below.

    ## Authentication vs authorization — module boundary

    Separate concerns, separate modules, enforced at different layers. The
    split is by **granularity**, not preference: the dividing question is
...

...

...
    user may do.

...

...

...

...
servers:
...
paths:
...
  /api/users/me:
    get:
      tags: [users]
      operationId: getCurrentUser
      summary: Return the current user's profile and role
      security:
        - sessionCookie: []
      x-auth-enforcement: GATEWAY
      description: |
        Profile and role for the caller identified by the gateway-injected
        `user_id`. Called alongside `GET /api/auth/session` on SPA bootstrap
        — see the module boundary table in `info.description`. This is the
        endpoint that v0.1.0's `GET /api/auth/me` was split into.
      responses:
        "200":
          description: Current user's profile.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/users:
    post:
      tags: [users]
      operationId: provisionUser
      summary: Provision a new user record (invite-first)
      security:
        - sessionCookie: []
      x-auth-enforcement: GATEWAY # authentication only
      x-authorization: |
        users module — delegation-scoped, see table below. Not a single
        require_role() call.
      description: |
        Creates a `users` row keyed on `email`, with `external_id`/`tenant_id`
        left NULL — first login binds them (see `/api/auth/callback` step 6).
        No session is created here; this is invitation, not authentication.

        `provisioned_by` carries the supervision edge for a
        `teaching_assistant` — see DECISION LOG [0.5.6]. It is derived, never
        taken at face value:

        | Caller       | `role`               | `supervisor_id` | `provisioned_by` |
        |--------------|----------------------|-----------------|------------------|
        | `root_admin` | `teaching_assistant` | an instructor   | that instructor  |
        | `root_admin` | `teaching_assistant` | omitted/null    | `NULL`           |
        | `root_admin` | `instructor`         | must be absent  | the admin        |
        | `instructor` | `teaching_assistant` | absent or self  | the instructor   |
        | `instructor` | `teaching_assistant` | a colleague     | `403`            |

        An admin-created assistant with no named instructor is genuinely
        unassigned — `NULL`, not the admin's id. The admin does not supervise,
        and an assistant with no supervisor cannot screen, so recording the
        admin there would grant access nobody granted.

        ### Delegation-scoped authorization
        Enforces the delegation chain from `info.description` §Provisioning:

        | Caller role           | May provision role     |
        |------------------------|-------------------------|
        | `root_admin`           | `instructor`             |
        | `instructor`           | `teaching_assistant`     |
        | `teaching_assistant`   | — (`403`)                |

        `root_admin` itself can never be a value of `role` in the request
        body — there is exactly one, created only by the seed migration,
        `provisioned_by = NULL`. Requesting it is `403` regardless of caller.

        > **OPEN QUESTION — confirm before this ships.** `teaching_assistant`
        > is scoped as an E5 extension in the proposal (not yet built in the
        > web app). This table gives `instructor` real delegation authority
        > to provision TA accounts *today*, matching the chain already
        > committed to in v0.3.0's decision log. If that authority shouldn't
        > be live until E5 actually exists, restrict the caller column to
        > `root_admin` only for now and revisit this table when E5 starts.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserCreateRequest" }
      responses:
        "201":
          description: User provisioned. `first_login_at` is null until they sign in.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: |
            Caller's role may not provision the requested role (see
            delegation table), or the request named `root_admin`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: |
            `email` already has a `users` row. If that row is soft-deleted,
            use `POST /api/users/{id}/restore` instead — restoring an
            existing account and provisioning a new one are different
            audited actions, and collapsing them into upsert semantics here
            would blur that distinction in the audit trail.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

    get:
      tags: [users]
      operationId: listUsers
      summary: List / filter users
      security:
        - sessionCookie: []
      x-auth-enforcement: GATEWAY
      x-authorization: "users module — delegation-scoped, see DECISION LOG [0.5.5]"
      description: |
        Administrative directory view. Delegation-scoped, not a flat
        `root_admin` gate — see DECISION LOG [0.5.5] for why this reverses
        [0.4.0]'s original root_admin-only text:
          - `root_admin` sees every row, any role, including other
            `root_admin` rows.
          - `instructor` sees only `teaching_assistant` rows they themselves
            provisioned. Never other instructors, never `root_admin`, never
            TAs provisioned by someone else.
          - `teaching_assistant` gets `403` — this is a role-level floor
            gate for TAs, not an object-level scoping question.

        `role` is a *filter* on an already-scoped result set, not an
        authorization input — same filter-vs-create asymmetry as elsewhere
        in this contract. For an `instructor` caller, the effective scope is
        always `role=teaching_assistant AND provisioned_by=<caller>`
        regardless of what `role` value is passed; requesting a `role` value
        outside that scope silently returns an empty page rather than
        `403`, so there's nothing to enumerate about whether e.g. another
        instructor's account exists.
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [root_admin, instructor, teaching_assistant]
        - name: include_deleted
          in: query
          schema: { type: boolean, default: false }
          description: Include soft-deleted rows. Excluded by default.
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
        - name: cursor
          in: query
          schema: { type: string }
          description: |
            Opaque. Base64 encoding of the last `id` from the previous
            page. Keyset-paginated ordered by `id` ascending.
      responses:
        "200":
          description: Paginated users.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/User" }
                  next_cursor:
                    type: string
                    nullable: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/users/{id}:
    get:
      tags: [users]
      operationId: getUser
      summary: Get a single user
      security:
        - sessionCookie: []
      x-auth-enforcement: GATEWAY
      x-authorization: "users module — object-level, see DECISION LOG [0.5.2]"
      description: |
        Object-scoped, not role-gated — every provisioned caller may hit
        this route, visibility of the *target* is what varies:
          - `root_admin` may view anyone, including other `root_admin` rows.
          - Anyone may view their own record.
          - `instructor` may view other `instructor` and `teaching_assistant`
            rows, but not `root_admin` rows.
          - `teaching_assistant` may view any `instructor` row, and any
            `teaching_assistant` row sharing the same `provisioned_by`
            (i.e. created by the same instructor), but not `root_admin` rows
            and not TAs provisioned by a different instructor.
        Out-of-scope targets return `404`, same as a nonexistent id — see
        the `NotFound` schema. There is no `403` path for this endpoint: it
        has no function-level role gate, so every denial is object-level.
        Functionally redundant with `GET /api/users/me` for the self case —
        kept anyway so an admin-UI table and a single-user drill-down can
        hit the same shape via the same route pattern.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: User record.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags: [users]
      operationId: deleteUser
      summary: Soft-delete a user and revoke all of their sessions
      security:
        - sessionCookie: []
      x-auth-enforcement: GATEWAY
      x-authorization: "users module — delegation-scoped, see DECISION LOG [0.5.4]"
      description: |
        Sets `deleted_at = now()`. Never a hard delete — the audit trail's
        referential integrity depends on the row surviving, same rationale
        as soft-deleted sessions.

        Delegation-scoped, same chain as `POST /api/users`: `root_admin` may
        delete anyone; `instructor` may only delete `teaching_assistant`
        rows they themselves provisioned; `teaching_assistant` may delete
        no one. Unauthorized attempts return `403`, not `404` — this is a
        role-level gate, not an object-level visibility check, so there's
        nothing to hide about the target's existence once the actor is
        already known to be able to see users of that kind.

        **Immediately revokes every active session belonging to this user**,
        not just the profile flag. This is the users module calling back
        into auth-owned state — the mirror image of auth calling
        `users.resolve_or_reject` at callback time (see
        `/api/auth/callback` step 6), just in the other direction:

        ```python
        await sessions.revoke_all_for_user(user_id)
        # internal cross-module call, not a public endpoint.
        ```

        This is why the gateway authorizer's `result_ttl_seconds = 0` matters
        (see `x-gateway-configuration`): without it, a cached "session valid"
        decision could outlive this revocation by up to the cache TTL, and
        "immediately" would be false.

        A deleted user hitting `/api/auth/callback` afterward is rejected
        with the distinct `deactivated` error code, not `not_provisioned` —
        the audit trail should be able to tell "never had access" apart from
        "access was revoked."

        Distinct from `DELETE /api/admin/sessions/{session_id}`, which
        revokes exactly one session and does not touch the user record.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: |
            Deleted; all sessions for this user revoked. Idempotent — an
            already-deleted user returns 204.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/users/{id}/supervisor:
    post:
      tags: [users]
      operationId: setUserSupervisor
      summary: Assign, move or unassign a teaching assistant's supervisor
      security:
        - sessionCookie: []
      x-auth-enforcement: GATEWAY
      x-authorization: "users module — see DECISION LOG [0.5.6]"
      description: |
        Sets `provisioned_by` on a `teaching_assistant`. Its own endpoint for
        the same reason `/role` is: supervision decides who may manage the row
        and whether the assistant can screen at all, so changing it is a
        distinct action rather than a field on a general update.

        Placing people is a `root_admin` act. An `instructor` may only
        **release** their own assistant — `supervisor_id: null` on a row whose
        `provisioned_by` is already theirs. Anything else from an instructor is
        `403`, including claiming an unassigned account: `provisioned_by` is
        what grants management of the row, so an instructor who could assign
        could hand a colleague access, or help themselves to someone else's
        assistant. A `teaching_assistant` reaches this endpoint never.

        Unassigning **revokes the assistant's live sessions**, in the same
        commit. The SPA answers "may this person screen?" from the session
        payload it already holds, so an assistant who kept a session would keep
        screening until they happened to reload.

        A move does not revoke: they still have a supervisor, so nothing they
        may do has changed.

        Setting the supervisor to the value it already holds is a no-op that
        returns `200` — no revocation, no audit noise.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SupervisorChangeRequest" }
      responses:
        "200":
          description: Updated user, so the caller sees the resulting assignment.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400":
          description: |
            The target is not a `teaching_assistant`, or the proposed
            supervisor is missing, not an `instructor`, or not `active`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: |
            The target is deleted. Deletion is terminal, so there is nothing to
            assign them to.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-pp_session
      description: |
        Opaque 256-bit CSPRNG session identifier. Carries no claims and is not
        a token — it is a primary key into the `sessions` table and is
        meaningless outside this system.

        Attributes (per `draft-ietf-oauth-browser-based-apps` §6.1.3.2):
        `Secure` and `HttpOnly` are MUST; `SameSite=Strict`, `Path=/`, absent
        `Domain`, and the `__Host-` prefix are SHOULD and all applied.

        Establishes identity only. Authorization is evaluated separately by the
        users module.
  schemas:
...
    User:
      type: object
      required: [id, email, role, created_at]
      description: |
        Profile and authorization state. Owned entirely by the users module —
        the auth module never reads or writes `role` or `provisioned_by`; see
        the module boundary table in `info.description`.
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
          description: |
            Provisioning key (CITEXT, case-insensitive). Authoritative before
            first login; display data only afterward — see "Provisioning" in
            `info.description`.
        display_name:
          type: string
          nullable: true
          description: NULL until first login; set from the Entra `name` claim thereafter.
        role:
          type: string
          enum: [root_admin, instructor, teaching_assistant]
        provisioned_by:
          type: string
          format: uuid
          nullable: true
          description: The provisioning caller's `user_id`. NULL only for the seeded `root_admin`.
        first_login_at:
          type: string
          format: date-time
          nullable: true
          description: NULL means invite pending — this row has never completed a login.
        created_at:
          type: string
          format: date-time
        deleted_at:
          type: string
          format: date-time
          nullable: true
          description: Soft-delete marker. NULL means active.
    UserCreateRequest:
      type: object
      required: [email, role]
      properties:
        email:
          type: string
          format: email
        role:
          type: string
          enum: [instructor, teaching_assistant]
          description: |
            `root_admin` is rejected here — see the delegation table on
            `POST /api/users`.
        display_name:
          type: string
          description: |
            Optional admin-UI placeholder shown before first login.
            Overwritten by the Entra `name` claim on first login — never
            treat this value as authoritative.
        supervisor_id:
          type: string
          format: uuid
          nullable: true
          description: |
            The instructor who will supervise a new `teaching_assistant`.
            `root_admin` only; an `instructor` may omit it or name themselves,
            and naming anyone else is `403`. Must be absent for any other
            `role`. Omitted or null from an admin means unassigned — see the
            derivation table on `POST /api/users`.
    SupervisorChangeRequest:
      type: object
      required: [supervisor_id]
      properties:
        supervisor_id:
          type: string
          format: uuid
          nullable: true
          description: |
            The instructor who will supervise this assistant, or null to
            unassign them. Null is spelled out rather than omitted: the caller
            is asking for a change, not accepting a default.
...
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Stable machine-readable code. Safe to branch on.
          enum:
            - invalid_request
            - unauthenticated # 401 — no valid session (auth, gateway)
            - session_expired # 401 — session existed, clock fired (auth)
            - not_provisioned # authenticated at Entra, no user record (users)
            - deactivated # authenticated at Entra, user record exists
              # but soft-deleted (users)
            - forbidden # 403 — valid session, insufficient role (users)
            - not_found
            - conflict # 409 — duplicate email, or invalid state
              # transition e.g. restoring a non-deleted user (users)
            - state_mismatch
            - upstream_idp_error
            - payload_too_large # 413 — answer_text over 10,000 chars (checks)
            - question_text_required # 400 — served detector scores on
              # (question, answer) pairs and question_text
              # was omitted. See GET /api/detector. (checks)
            - detector_unavailable # 503 — detector service down (checks)
            - detector_timeout # 504 — detector exceeded 10s budget (checks)
            - internal_error
        message:
          type: string
          description: |
            Human-readable. Never interpolates unsanitised upstream input
            (notably Entra's `error_description`).
        request_id:
          type: string
          description: Correlates with CloudWatch logs and the audit trail.
  responses:
    BadRequest:
      description: Malformed or invalid request.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: |
        No session, or the session is expired or revoked. Emitted by the
        **gateway authorizer**, before application code runs.

        API Gateway renders these from its own gateway-response templates, not
        from FastAPI. That template MUST be configured to emit the `Error`
        schema — otherwise the SPA receives two structurally different 401
        bodies depending on which layer rejected the request, and the frontend
        ends up parsing both defensively. Fix in Terraform.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: |
        Valid session, but the user's role does not permit this operation.
        Emitted by the **users module**, inside application code — never by the
        gateway. See DECISION LOG [0.3.0].
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...